gh-422: Upgrade storybook and reorganize stories
