+++ .prettierignore
@@ -3,4 +3,5 @@ dist/
 lib/
 node_modules/
 .cache/
+temp
 *.(jpeg|png|gif)
